Numberous more problems with the circuit design and PCB layout errors found and corrected as a result of debugging and making the RF section work, today, finally.
There remains and issue with the SPI port disable jumper not working as it should -- I think.  I got around this by fluke, when I realised that holding down the reset button (somehow) enables my AVR ISP mkII to program the AVR chip, despite the suspected SPI bus conflict between the ISP and level shofter chip. (Not that any of that makes sense to me right now. But that's what seemed to be the case, at the time. I'll sort it later.)